requirements.txt
aiomysql==0.2.0
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
certifi==2025.6.15
cffi==1.17.1
click==8.2.1
cryptography==45.0.5
distro==1.9.0
ecdsa==0.19.1
fastapi==0.116.0
greenlet==3.2.3
grpcio==1.73.1
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
jiter==0.10.0
numpy==2.3.1
openai==1.93.1
packaging==25.0
passlib==1.7.4
pluggy==1.6.0
portalocker==2.10.1
protobuf==6.31.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyMySQL==1.1.1
pytest==8.4.1
pytest-order==1.3.0
python-dotenv==1.1.1
python-jose==3.5.0
qdrant-client==1.14.3
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
